% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAB.R
\name{JAB01}
\alias{JAB01}
\title{JAB01: Compute the Jeffreys Approximate Objective Bayes factor from p-value}
\usage{
JAB01(N, pVal, model, R = NULL, C = NULL, I = NULL)
}
\arguments{
\item{N}{Sample size (number of observations).}

\item{pVal}{P-value.}

\item{model}{The statistical model being used, one of:
\itemize{
\item "t-test"
\item "linear_regression"
\item "logistic_regression"
\item "cox"
\item "anova" (one-way ANOVA)
\item "repeated_measures"
\item "chi_squared"
}}

\item{R}{(Optional) Number of rows for chi-squared tests.}

\item{C}{(Optional) Number of columns for chi-squared tests.}

\item{I}{(Optional) Number of conditions for one-way ANOVA and repeated measures.}
}
\value{
A numeric value corresponding to the Jeffreys Approximate Bayes factor.
}
\description{
This function calculates the Jeffreys Approximate Objective Bayes (JAB) factor
based on the formula:
N^(k/2) * exp(-0.5 * (N - 1)/N * Q_k(1 - p-value)).
}
\examples{
JAB01(100, 0.05, "t-test")
JAB01(200, 0.01, "chi_squared", R = 3, C = 4)
}
